+++ crates/ewwii/src/widgets/build_widget.rs
@@ -60,7 +60,7 @@ fn build_gtk_widget_from_node(
             build_circular_progress_bar(props, widget_reg)?.upcast()
         }
         WidgetNode::GtkUI { props } => build_gtk_ui_file(props)?.upcast(),
-        // WidgetNode::Graph { props } => build_graph(props, widget_reg)?.upcast(),
+        WidgetNode::Graph { props } => build_graph(props, widget_reg)?.upcast(),
         // WidgetNode::Transform { props } => build_transform(props, widget_reg)?.upcast(),
         WidgetNode::Scale { props } => build_gtk_scale(props, widget_reg)?.upcast(),
         WidgetNode::Progress { props } => build_gtk_progress(props, widget_reg)?.upcast(),